% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmc-package.R
\docType{package}
\name{SVMC-package}
\alias{SVMC}
\alias{SVMC-package}
\title{SVMC: Structural Variant Mapping and Characterization for Bacterial Genomes}
\description{
A pipeline and analysis framework for identifying, classifying, and validating large-scale structural variation in complete bacterial genome assemblies. SVMC orients genomes to the replication origin (OriC) using multi-marker scoring, aligns genomes with MUMmer, classifies structural variant classes including replication-associated structural rearrangements (RASRs), groups recurrent variants into motifs, and identifies lineage-defining motifs using a phi-coefficient exclusivity score validated against MASH distance clustering.
}
\author{
\strong{Maintainer}: Matthew D'Iorio \email{matthew.diorio@mail.mcgill.ca}

}
\keyword{internal}
